+++ src/app/shared/components/sui-dropdown/sui-multi-select-search-dropdown.component.ts
@@ -100,10 +100,23 @@ export class SuiMultiSelectSearchDropdownComponent implements ControlValueAccess
     const options: object = {...this.options,...{
         onChange: (value:string | string[]) => {
           // avoid firing useless change events:
-          // - on clearing the dropdown list
           // - values did not change
           // - in single selection mode: empty field not allows
+
           let shouldFire = !this.suppressChange  && value != undefined;
+          const isCleared =
+            value === undefined ||
+            (typeof value === 'string' && value.trim() === '') ||
+            (Array.isArray(value) && value.length === 0);
+
+          if (isCleared && shouldFire) {
+            if (this.value !== null) {
+              this.value = "";
+              this.onChange("");
+            }
+            return;
+          }
+
           if (this.multiple){
             value = value as string[];
             this.value = this.value as string[];